logout works for redux and session
