+++ backend/passport/logout.js
@@ -4,14 +4,20 @@ const express = require('express');
 var router = express.Router();
 
 module.exports = (passport) => {
-		 router.post('/logout', (req, res) => {
-   console.log('tf is req', req, 'wb data', req.data, 'wb body', req.body);
-		 req.logout();
- });
+  router.post('/logout', (req, res) => {
+    console.log('session before', req.session);
+    req.logout();
+    console.log('session after', req.session);
+    // TODO: Need to understand if I need to destroy session totally or just destroy passport portion
+    // req.session.destroy((err) => {
+    //   if (err) {
+    //     console.log('error logging out', err);
+    //   }
+    //   res.send(true);
+    // });
+    res.send(true);
+    return true;
+  });
+
 		 return router;
 };
-
-    // Generates hash using bCrypt, storing password safely
-	    var createHash = (password) => {
-	        return bCrypt.hashSync(password, bCrypt.genSaltSync(10), null);
-	    };
